+++ rust/operator-binary/src/crd/catalog/mod.rs
@@ -49,10 +49,19 @@ pub mod versioned {
         /// The `connector` defines which connector is used.
         pub connector: TrinoCatalogConnector,
 
-        #[serde(default)]
         /// The `configOverrides` allow overriding arbitrary Trino settings.
         /// For example, for Hive you could add `hive.metastore.username: trino`.
+        #[serde(default)]
         pub config_overrides: HashMap<String, String>,
+
+        /// List of config properties which should be removed.
+        ///
+        /// This is helpful, because Trino fails to start in case you have any unused config
+        /// properties. The removals are executed after the `configOverrides`.
+        ///
+        /// This field is experimental, and might be replaced by a more generic mechanism to edit config properties
+        #[serde(default, rename = "experimentalConfigRemovals")]
+        pub config_removals: Vec<String>,
     }
 }