+++ CHANGELOG.md
@@ -12,6 +12,8 @@ All notable changes to this project will be documented in this file.
 - Support configuring JVM arguments ([#677]).
 - Aggregate emitted Kubernetes events on the CustomResources ([#677]).
 - Support for Trino 470 ([#705]).
+- Support removing properties from catalogs.
+  This is helpful, because Trino fails to start in case you have any unused config properties ([#713]).
 - Support `access-control.properties` in configOverrides ([#721]).
 
 ### Changed
@@ -38,6 +40,7 @@ All notable changes to this project will be documented in this file.
 [#694]: https://github.com/stackabletech/trino-operator/pull/694
 [#695]: https://github.com/stackabletech/trino-operator/pull/695
 [#705]: https://github.com/stackabletech/trino-operator/pull/705
+[#713]: https://github.com/stackabletech/trino-operator/pull/713
 [#715]: https://github.com/stackabletech/trino-operator/pull/715
 [#717]: https://github.com/stackabletech/trino-operator/pull/717
 [#721]: https://github.com/stackabletech/trino-operator/pull/721